requirements.txt
streamlit
groq
pandas
python-dotenv
requests
